front-end and browser config

+++ src/contracts/EthSwap.sol
@@ -2,29 +2,30 @@
 pragma solidity ^0.5.0;
 
 import "./Token.sol";
- 
 
-contract EthSwap{
+contract EthSwap {
     // state variables, stored in the blockchain, can use as getter
     string public name = "EthSwap Instant Exchange";
     Token public token;
-    uint public rate = 100;
+    uint256 public rate = 100;
 
     event TokenPurchased(
         address account,
         address token,
-        uint amount,
-        uint rate
+        uint256 amount,
+        uint256 rate
     );
+
     // param: token.address
     constructor(Token _token) public {
         token = _token;
     }
+
     // transfer tokens to buyer
     // msg is a global var in solidity, sender will be the person call the function
     function buyTokens() public payable {
         // calc the number of tokens to buy
-        uint tokenAmount = msg.value * rate;
+        uint256 tokenAmount = msg.value * rate;
         // amount of Eth *Redemption rate(# of tokens they receive of 1 ether)
 
         // require the EthSwap has enouth tokens
@@ -36,14 +37,13 @@ contract EthSwap{
         emit TokenPurchased(msg.sender, address(token), tokenAmount, rate);
     }
 
-    function sellTokens(uint _amount) public {
+    function sellTokens(uint256 _amount) public {
         // calculate the amount of Ether to redeem
-        uint etherAmount = _amount / rate;
+        uint256 etherAmount = _amount / rate;
         // maker sure the EtherSwap greter than amount
         require(address(this).balance >= etherAmount);
         // perform sale operation
         token.transferFrom(msg.sender, address(this), _amount);
         msg.sender.transfer(etherAmount);
-
     }
 }
